예측 확률 버그 수정

+++ Common Library/DotNet_Framework/PredictorAPI/Models/Football/FootballTeamStat.cs
@@ -22,6 +22,6 @@ namespace PredictorAPI.Models.Football
         public double DefTrend { get; set; }
 
         [JsonProperty("rest_date")]
-        public int RestDate { get; set; }
+        public double RestDate { get; set; }
     }
 }
